+++ consts.go
@@ -23,8 +23,6 @@ CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 
 package firebirdsql
 
-import "database/sql/driver"
-
 const (
 	ISC_TIME_SECONDS_PRECISION = 10000
 
@@ -433,7 +431,7 @@ const (
 )
 
 const (
-	ISOLATION_LEVEL_READ_COMMITED_LEGACY driver.IsolationLevel = iota
+	ISOLATION_LEVEL_READ_COMMITED_LEGACY = iota
 	ISOLATION_LEVEL_READ_COMMITED
 	ISOLATION_LEVEL_REPEATABLE_READ
 	ISOLATION_LEVEL_SERIALIZABLE
